refactor nwm config class into module
